refactor: remove GainProcessor redundancy, pre-allocate STFT block buffer, fix raw pointer pattern
- Delete GainProcessor (logic absorbed directly into GainStage)
- Pre-allocate block_ member in WindowedFFTStage to avoid per-frame heap allocation
- Use make_unique + .get() in WavProcessor before transferring ownership to Pipeline

+++ include/audio/GainStage.hpp
@@ -8,5 +8,5 @@ class GainStage : public AudioStage {
     void process(std::vector<float> &samples, uint16_t numChannels) override;
 
   private:
-    double gain_;
+    float gain_;
 };
